Pass display info into ShellUI when in Enterprise

+++ src/Backend/DisplaysManager/DisplaysManager.cpp
@@ -49,7 +49,7 @@ void DisplaysManager::HandleDisplayDetected(const DisplayInfo& info)
 
 // This will be called before the individual DisplayInfo objects are passed in from
 // the WindowsAPI.
-void DisplaysManager::HandleNumberOfDisplaysChanged(uint8_t numDisplays)
+void DisplaysManager::HandleNumberOfDisplaysChanged(unsigned int numDisplays)
 {
    NumDisplays = numDisplays;
    CurrentEvent = DisplayConfigEvent();
